+++ src/Mapbender/ManagerBundle/Controller/ElementController.php
@@ -15,6 +15,8 @@ use FOM\ManagerBundle\Configuration\Route as ManagerRoute;
 use Mapbender\CoreBundle\Entity\Element;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Acl\Domain\ObjectIdentity;
+use Symfony\Component\Security\Acl\Permission\MaskBuilder;
 
 /**
  * Class ElementController
@@ -179,10 +181,13 @@ class ElementController extends ApplicationControllerBase
         $entityManager->detach($element); // prevent element from being stored with default config/stored again
 
         $application = $this->requireApplication($slug);
-        $form = $this->createForm('FOM\UserBundle\Form\Type\ACLType', $element, array(
-            'create_standard_permissions' => false,
-            'permissions' => array(
-                1 => 'View',
+        $form = $this->createForm('Symfony\Component\Form\Extension\Core\Type\FormType', null, array(
+            'label' => false,
+        ));
+        $form->add('acl', 'FOM\UserBundle\Form\Type\ACLType', array(
+            'object_identity' => ObjectIdentity::fromDomainObject($element),
+            'entry_options' => array(
+                'mask' => MaskBuilder::MASK_VIEW,
             ),
         ));
         $form->handleRequest($request);
@@ -191,7 +196,7 @@ class ElementController extends ApplicationControllerBase
             try {
                 $application->setUpdated(new \DateTime('now'));
                 $entityManager->persist($application);
-                $this->getAclManager()->setObjectACEs($element, $form->get('ace')->getData());
+                $this->getAclManager()->setObjectACEs($element, $form->get('acl')->getData());
                 $entityManager->flush();
                 $entityManager->commit();
                 $this->addFlash('success', "Your element's access has been changed.");
@@ -202,6 +207,7 @@ class ElementController extends ApplicationControllerBase
             }
             return $this->redirectToRoute('mapbender_manager_application_edit', array(
                 'slug' => $slug,
+                '_fragment' => 'tabLayout',
             ));
         }
         return $this->render('@MapbenderManager/Element/security.html.twig', array(